feat: 메인 페이지 필터링 기능 추가
