+++ index.html
@@ -28,14 +28,14 @@
           id="category-assignment"
           class="category category-assignment"
         >
-          <option value="all" selected>과제</option>
+          <option value="" selected>과제</option>
         </select>
         <select
           name="category-class"
           id="category-class"
           class="category category-class"
         >
-          <option value="all" selected>반</option>
+          <option value="" selected>반</option>
           <option value="1">1-1</option>
           <option value="2">1-2</option>
           <option value="3">1-3</option>
@@ -46,8 +46,8 @@
           id="category-sort"
           class="category category-sort"
         >
-          <option value="latest" selected>최신순</option>
-          <option value="popular">인기순</option>
+          <option value="createdAt" selected>최신순</option>
+          <option value="likes">인기순</option>
         </select>
       </div>
     </nav>
