+++ src/js/common.js
@@ -1,3 +1,5 @@
+import hljs from "highlight.js";
+
 let elements = {};
 
 document.addEventListener("DOMContentLoaded", async () => {
@@ -152,3 +154,45 @@ export const setupModalControls = () => {
     }
   });
 };
+
+
+export const loadCodes = async (criteria = "createdAt", classNo = "", assignmentId= "") => {
+  const list = document.getElementById("post-list-container");
+  const codes = await customFetch(`/codes?criteria=${criteria}&classNo=${classNo}&assignmentId=${assignmentId}`);
+
+  if (codes.ok && codes.data) {
+    list.innerHTML = "";
+    for (let i = 0; i < codes.data.codes.length; i++) {
+      const code = codes.data.codes[i];
+      const article = document.createElement("article");
+      const formattedStudentNumber = String(code.user.studentNumber).padStart(
+          2,
+          "0"
+      );
+      article.className = "post-list-box";
+      article.innerHTML = `
+                <div class="code-text">
+                <a href="/code?id=${code.id}" class="detail-page">
+                  <pre><code class="language-python">${code.code}</code></pre>
+                </div>
+                </a>
+                <div class="code-information">
+                  <div class="info-box">
+                    <a href="/code?id=${code.id}" class="detail-page">
+                      <p class="title">${code.title}</p>
+                    </a>
+                    <a href="/user?id=${code.user.id}" class="detail-page">
+                      <p class="student-info">${code.user.grade}${code.user.classNumber}${formattedStudentNumber}${code.user.name}(${code.user.username})</p>
+                    </a>
+                  </div>
+                  <div class="like-box">
+                    <img src="/like.svg" alt="likeIcon" width="23px" />
+                    <p className="like-count" id="code-likes">${code.likes}</p>
+                  </div>
+                </div>
+                `;
+      list.appendChild(article);
+    }
+    hljs.highlightAll();
+  }
+}
